fixed example data creation

+++ src/testing/TestingDataCreation.cu
@@ -16,8 +16,8 @@ std::unique_ptr<__half[]> CreateSineSuperpostion(int amount_of_timesamples,
   for(int i=0; i<amount_of_timesamples; i++){
     float tmp = 0;
     for(int j=0; j<static_cast<int>(weights.size()); j++){
-      tmp += (weights[j] *
-              sin(2 * M_PI * (j + 1) * (i / amount_of_timesamples)));
+      tmp += (weights[j] * sin(2 * M_PI * (j + 1) *
+             (static_cast<double>(i) / amount_of_timesamples)));
     }
     data[i] = __float2half(tmp);
     data[i + amount_of_timesamples] = 0;
